# Optimize `max_boolean` by operating on u64 chunks