# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

# Rationale for this change

Vectorizing boolean operations using u64 chunks is generally more efficient.

<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

# What changes are included in this PR?

- Adds benchmarks for boolean min/max
- Operate on bit chunks instead of individual booleans, which can lead to massive speedups while not regressing the short-circuiting behavior of the existing implementation.

`cargo bench --bench aggregate_kernels -- "bool/max"` shows throughput improvements between 50% to 23390% on my machine.

<details><summary>Full benchmark results:</summary>

```
cargo bench --bench aggregate_kernels -- "bool/max"
bool/max nonnull mixed  time:   [4.9617 ns 4.9934 ns 5.0338 ns]
                        thrpt:  [ 13019 Gelem/s  13125 Gelem/s  13208 Gelem/s]
                 change:
                        time:   [-36.288% -34.043% -31.678%] (p = 0.00 < 0.05)
                        thrpt:  [+46.365% +51.614% +56.957%]
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) high mild
  8 (8.00%) high severe
bool/max nonnull false  time:   [565.07 ns 588.68 ns 611.85 ns]
                        thrpt:  [107.11 Gelem/s 111.33 Gelem/s 115.98 Gelem/s]
                 change:
                        time:   [-98.893% -98.849% -98.802%] (p = 0.00 < 0.05)
                        thrpt:  [+8243.8% +8589.5% +8936.8%]
                        Performance has improved.
bool/max nonnull true   time:   [5.0621 ns 5.0978 ns 5.1435 ns]
                        thrpt:  [ 12741 Gelem/s  12856 Gelem/s  12946 Gelem/s]
                 change:
                        time:   [-4.7786% -1.8889% +1.1654%] (p = 0.23 > 0.05)
                        thrpt:  [-1.1520% +1.9253% +5.0184%]
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe
bool/max nullable mixed time:   [9.4018 ns 9.5263 ns 9.6842 ns]
                        thrpt:  [6767.3 Gelem/s 6879.5 Gelem/s 6970.6 Gelem/s]
                 change:
                        time:   [-41.198% -38.405% -35.720%] (p = 0.00 < 0.05)
                        thrpt:  [+55.568% +62.350% +70.063%]
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  3 (3.00%) high mild
  10 (10.00%) high severe
bool/max nullable false time:   [1.1109 µs 1.1591 µs 1.2086 µs]
                        thrpt:  [54.223 Gelem/s 56.539 Gelem/s 58.994 Gelem/s]
                 change:
                        time:   [-99.592% -99.574% -99.556%] (p = 0.00 < 0.05)
                        thrpt:  [+22422% +23392% +24407%]
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
bool/max nullable true  time:   [9.0433 ns 9.0830 ns 9.1284 ns]
                        thrpt:  [7179.4 Gelem/s 7215.3 Gelem/s 7246.9 Gelem/s]
                 change:
                        time:   [-39.120% -37.104% -34.921%] (p = 0.00 < 0.05)
                        thrpt:  [+53.659% +58.992% +64.257%]
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) high mild
  10 (10.00%) high severe
```

</details>

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?

Faster `max_boolean`.

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->